Renaming LatticeBoltzmann to LB for simplicity
